requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
transformers
torch